refactor: 패키지명 변경
- model 패키지명을 domain으로 변경하였습니다.
