+++ src/main/java/domain/Lottos.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import dto.LottosResponse;
