+++ src/main/java/domain/RankType.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import dto.LottoRankDetailResponse;
