+++ src/test/java/domain/InvestmentOutcomeTest.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
