+++ src/test/java/domain/LottoMachineTest.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
