+++ src/test/java/domain/RankTypeTest.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
